[WRFE-15](refactor): import정리, log제거 등
